Move hreflang and canonical middleware to its own file
